Added event listeners except color and path
Removed border line on ellipse (bad OF handling)

Changed unsigned int to float in line->border

LoadModel now set the path variable if the model loaded succesfully

Added names to unanmed gameobject, and changed paths into bin/data

Added event listeners and action (except color and path)

+++ src/components/Line.cpp
@@ -20,11 +20,11 @@ ofColor Components::Line::getLineColor() {
     return this->lineColor;
 }
 
-unsigned int Components::Line::getLineWidth() {
+float Components::Line::getLineWidth() {
     return this->lineWidth;
 }
 
-unsigned int Components::Line::getLineLength() {
+float Components::Line::getLineLength() {
     return this->lineLength;
 }
 
@@ -32,11 +32,11 @@ void Components::Line::setLineColor(ofColor color) {
     this->lineColor = color;
 }
 
-void Components::Line::setLineWidth(unsigned int width) {
+void Components::Line::setLineWidth(float width) {
     this->lineWidth = width;
 }
 
-void Components::Line::setLineLength(unsigned int length) {
+void Components::Line::setLineLength(float length) {
     this->lineLength = length;
 }
